refactor: events moved to separate folder
